Allow math expressions in AccountWidget balance fields.
Extract shared math input column logic so account and card dashboard widgets resolve basic arithmetic before saving.
