+++ app/Filament/Widgets/CardWidget.php
@@ -2,12 +2,10 @@
 
 namespace App\Filament\Widgets;
 
+use App\Filament\Widgets\Concerns\HasMathInputColumn;
 use App\Models\Card;
-use App\Rules\ValidMathExpression;
-use App\Support\MathExpression;
 use Filament\Tables\Columns\Summarizers\Sum;
 use Filament\Tables\Columns\TextColumn;
-use Filament\Tables\Columns\TextInputColumn;
 use Filament\Tables\Table;
 use Filament\Widgets\TableWidget as BaseWidget;
 use Illuminate\Database\Eloquent\Model;
@@ -15,6 +13,8 @@ use Illuminate\Database\Query\Builder;
 
 class CardWidget extends BaseWidget
 {
+    use HasMathInputColumn;
+
     protected int|string|array $columnSpan = 'full';
 
     protected static ?string $heading = 'Cards';
@@ -60,18 +60,5 @@ class CardWidget extends BaseWidget
                     ->label('Pts Bal')
                     ->summarize(Sum::make()->numeric()->label('')),
             ]);
-    }
-
-    private function mathInputColumn(string $field, bool $asInteger = false): TextInputColumn
-    {
-        return TextInputColumn::make($field)
-            ->rules([new ValidMathExpression])
-            ->updateStateUsing(function ($record, $state) use ($field, $asInteger) {
-                $resolved = MathExpression::resolve($state);
-                $value = $asInteger ? (int) round($resolved) : $resolved;
-                $record->update([$field => $value]);
-
-                return $value;
-            });
     }
 }
